Use IDs for each model
This is much clearer modeling for a database and webapp than the old approach of using an Arc for the Note<->Card relationship

I'm still not sure the right way to model creating cards from a note because the database will need to give the Card IDs to use. I suspect the handler will do this. First, create a Note in the DB, and then create 1-2 Cards in the DB.